Model change and test, fab task
Changed the CaseXML model to include a field for styled XML, and a method to update it.
Also added a test for that stuff.
Added a few arguments to the generate styled XML function to specify skipping dups
without raising an exception, and also appended a comment to the generated XML which
says if the text didn't totally match up with the ALTO.
Added a fab task which loops through each volume, synchronously, and asyncrhonously
updates the styled text for all of the cases in that volume.