Correct chronological order of "cited by" page
If an opinion is cited by more than one page's worth of opinions, they are split across multiple pages and the button to go to the second page is currently labeled "older entries." That's backwards, since the citations are listed from oldest to newest. This PR swaps the "older entries" and "newer entries" labels to reflect that.