LABS setting change
The LABS boolean setting is now the LABS_HIDDEN list. In the labs urls.py, add a check to see if the project name is in the LABS_HIDDEN list. If we want to retire something or it needs to be taken out of production because it's problematic for some reason, just add it to the list in the base settings.